+++ client/tc/azext_tc/_params.py
@@ -14,10 +14,12 @@ from ._validators import (
     project_name_validator, project_name_or_id_validator, user_name_or_id_validator,
     tracking_id_validator, project_type_id_validator, project_type_id_validator_name, provider_id_validator,
     subscriptions_list_validator, provider_event_list_validator, url_validator, base_url_validator,
-    teamcloud_source_version_validator, providers_source_version_validator, auth_code_validator, properties_validator)
+    teamcloud_source_version_validator, providers_source_version_validator, auth_code_validator,
+    properties_validator, index_url_validator)
 
 from ._completers import (
-    get_project_completion_list, get_project_type_completion_list, get_provider_completion_list)
+    get_project_completion_list, get_project_type_completion_list, get_provider_completion_list,
+    get_provider_index_completion_list)
 
 from ._actions import CreateProviderReference
 
@@ -63,7 +65,8 @@ def load_arguments(self, _):
 
     for scope in ['tc deploy', 'tc upgrade', 'tc provider deploy', 'tc provider upgrade']:
         with self.argument_context(scope) as c:
-            c.argument('index_url', help='URL to custom index.json file.')
+            c.argument('index_url', help='URL to custom index.json file.',
+                       validator=index_url_validator)
 
     # Tags
 
@@ -199,7 +202,7 @@ def load_arguments(self, _):
                    type=str, help='Provider id.',
                    validator=provider_id_validator)
 
-    for scope in ['tc provider show', 'tc provider delete']:
+    for scope in ['tc provider show', 'tc provider delete', 'tc provider upgrade']:
         with self.argument_context(scope) as c:
             c.argument('provider', options_list=['--name', '-n'],
                        type=str, help='Provider id.',
@@ -212,6 +215,8 @@ def load_arguments(self, _):
                        help='Space-seperated provider ids.',
                        validator=provider_event_list_validator)
             c.argument('properties', properties_type)
+            # c.argument('command_mode', get_enum_type(
+            #     ['Simple', 'Extended'], default='Simple'), help='Provider command mode.')
 
     with self.argument_context('tc provider create') as c:
         c.argument('url', type=str, help='Provider url.',
@@ -221,16 +226,19 @@ def load_arguments(self, _):
 
     for scope in ['tc provider deploy', 'tc provider upgrade']:
         with self.argument_context(scope) as c:
-            c.argument('provider', get_enum_type(['azure.appinsights', 'azure.devops', 'azure.devtestlabs']),
-                       options_list=['--name', '-n'], help='Provider id.')
             c.argument('version', options_list=['--version', '-v'],
                        type=str, help='Provider version. Default: latest stable.',
                        validator=providers_source_version_validator)
             c.argument('prerelease', options_list=['--pre'], action='store_true',
                        help='Deploy latest prerelease version.')
-            c.argument('index_url', help='URL to custom index.json file.')
+            c.argument('index_url', help='URL to custom index.json file.',
+                       validator=index_url_validator)
 
     with self.argument_context('tc provider deploy') as c:
+        c.argument('provider', options_list=['--name', '-n'],
+                   type=str, help='Provider id.',
+                   validator=provider_id_validator,
+                   completer=get_provider_index_completion_list)
         c.argument('resource_group_name', resource_group_name_type,
                    help='Name of resource group.')
         c.argument('location', get_location_type(self.cli_ctx))
@@ -239,4 +247,5 @@ def load_arguments(self, _):
     with self.argument_context('tc provider list-available') as c:
         c.argument('show_details', action='store_true', options_list=['--show-details', '-d'],
                    help='Show the raw data from the providers index.')
-        c.argument('index_url', help='URL to custom index.json file.')
+        c.argument('index_url', help='URL to custom index.json file.',
+                   validator=index_url_validator)