+++ src/TeamCloud.Orchestrator/Activities/ProjectListActivity.cs
@@ -11,6 +11,7 @@ using Microsoft.Azure.WebJobs;
 using Microsoft.Azure.WebJobs.Extensions.DurableTask;
 using TeamCloud.Data;
 using TeamCloud.Model.Data;
+using TeamCloud.Orchestration;
 
 namespace TeamCloud.Orchestrator.Activities
 {
@@ -37,5 +38,37 @@ namespace TeamCloud.Orchestrator.Activities
                 .ToListAsync()
                 .ConfigureAwait(false);
         }
+    }
+
+    public class ProjectListByIdActivity
+    {
+        private readonly IProjectsRepository projectsRepository;
+
+        public ProjectListByIdActivity(IProjectsRepository projectsRepository)
+        {
+            this.projectsRepository = projectsRepository ?? throw new ArgumentNullException(nameof(projectsRepository));
+        }
+
+        [FunctionName(nameof(ProjectListByIdActivity))]
+        public async Task<IEnumerable<Project>> RunActivity(
+            [ActivityTrigger] IList<string> projectIds)
+        {
+            var projects = projectsRepository
+                .ListAsync(projectIds);
+
+            return await projects
+                .ToListAsync()
+                .ConfigureAwait(false);
+        }
+    }
+
+
+    internal static class ProjectListExtension
+    {
+        public static Task<IEnumerable<Project>> ListProjectsAsync(this IDurableOrchestrationContext durableOrchestrationContext)
+            => durableOrchestrationContext.CallActivityWithRetryAsync<IEnumerable<Project>>(nameof(ProjectListActivity), null);
+
+        public static Task<IEnumerable<Project>> ListProjectsAsync(this IDurableOrchestrationContext durableOrchestrationContext, IList<string> projectIds)
+            => durableOrchestrationContext.CallActivityWithRetryAsync<IEnumerable<Project>>(nameof(ProjectListByIdActivity), projectIds);
     }
 }