+++ src/TeamCloud.Orchestrator/Activities/ProjectSetActivity.cs
@@ -9,6 +9,8 @@ using Microsoft.Azure.WebJobs;
 using Microsoft.Azure.WebJobs.Extensions.DurableTask;
 using TeamCloud.Data;
 using TeamCloud.Model.Data;
+using TeamCloud.Orchestration;
+using TeamCloud.Orchestrator.Entities;
 
 namespace TeamCloud.Orchestrator.Activities
 {
@@ -34,5 +36,13 @@ namespace TeamCloud.Orchestrator.Activities
 
             return newProject;
         }
+    }
+
+    internal static class ProjectSetExtension
+    {
+        public static Task<Project> SetProjectAsync(this IDurableOrchestrationContext functionContext, Project project, bool allowUnsafe = false)
+            => functionContext.IsLockedByContainerDocument(project) || allowUnsafe
+            ? functionContext.CallActivityWithRetryAsync<Project>(nameof(ProjectSetActivity), project)
+            : throw new NotSupportedException($"Unable to set project '{project.Id}' without acquired lock");
     }
 }