+++ src/TeamCloud.Orchestrator/Activities/UserDeleteActivity.cs
@@ -9,6 +9,8 @@ using Microsoft.Azure.WebJobs;
 using Microsoft.Azure.WebJobs.Extensions.DurableTask;
 using TeamCloud.Data;
 using TeamCloud.Model.Data;
+using TeamCloud.Orchestration;
+using TeamCloud.Orchestrator.Entities;
 
 namespace TeamCloud.Orchestrator.Activities
 {
@@ -32,5 +34,16 @@ namespace TeamCloud.Orchestrator.Activities
                 .RemoveAsync(user)
                 .ConfigureAwait(false);
         }
+    }
+
+    internal static class UserDeleteExtension
+    {
+        public static Task<User> DeleteUserAsync(this IDurableOrchestrationContext functionContext, string userId, bool allowUnsafe = false)
+            => DeleteUserAsync(functionContext, Guid.Parse(userId), allowUnsafe);
+
+        public static Task<User> DeleteUserAsync(this IDurableOrchestrationContext functionContext, Guid userId, bool allowUnsafe = false)
+            => functionContext.IsLockedBy<User>(userId.ToString()) || allowUnsafe
+            ? functionContext.CallActivityWithRetryAsync<User>(nameof(UserDeleteActivity), userId)
+            : throw new NotSupportedException($"Unable to delete user '{userId}' without acquired lock");
     }
 }