+++ src/TeamCloud.Orchestrator/Activities/UserProjectMembershipDeleteActivity.cs
@@ -9,6 +9,8 @@ using Microsoft.Azure.WebJobs;
 using Microsoft.Azure.WebJobs.Extensions.DurableTask;
 using TeamCloud.Data;
 using TeamCloud.Model.Data;
+using TeamCloud.Orchestration;
+using TeamCloud.Orchestrator.Entities;
 
 namespace TeamCloud.Orchestrator.Activities
 {
@@ -36,5 +38,13 @@ namespace TeamCloud.Orchestrator.Activities
 
             return updatedUser;
         }
+    }
+
+    internal static class UserProjectMembershipDeleteExtension
+    {
+        public static Task<User> DeleteUserProjectMembershipAsync(this IDurableOrchestrationContext functionContext, User user, string projectId, bool allowUnsafe = false)
+            => functionContext.IsLockedBy<User>(user.Id) || allowUnsafe
+            ? functionContext.CallActivityWithRetryAsync<User>(nameof(UserProjectMembershipDeleteActivity), (user, projectId))
+            : throw new NotSupportedException($"Unable to delete project membership without acquired for user {user.Id} lock");
     }
 }