+++ src/TeamCloud.Orchestrator/Activities/UserProjectMembershipSetActivity.cs
@@ -9,6 +9,8 @@ using Microsoft.Azure.WebJobs;
 using Microsoft.Azure.WebJobs.Extensions.DurableTask;
 using TeamCloud.Data;
 using TeamCloud.Model.Data;
+using TeamCloud.Orchestration;
+using TeamCloud.Orchestrator.Entities;
 
 namespace TeamCloud.Orchestrator.Activities
 {
@@ -47,5 +49,13 @@ namespace TeamCloud.Orchestrator.Activities
 
             return user;
         }
+    }
+
+    internal static class UserProjectMembershipSetExtension
+    {
+        public static Task<User> SetUserProjectMembershipAsync(this IDurableOrchestrationContext functionContext, User user, string projectId, bool allowUnsafe = false)
+            => functionContext.IsLockedBy<User>(user.Id) || allowUnsafe
+            ? functionContext.CallActivityWithRetryAsync<User>(nameof(UserProjectMembershipSetActivity), (user, projectId))
+            : throw new NotSupportedException($"Unable to create or update project membership without acquired for user {user.Id} lock");
     }
 }