+++ client/tc/HISTORY.rst
@@ -3,6 +3,24 @@
 Release History
 ===============
 
+0.4.4
+++++++
+* Preview release
+* az project-type create: Fix provider id validation bug
+
+0.4.3
+++++++
+* Preview release
+* Fix bug in tc upgrade
+* Fix bug in user table output
+
+0.4.2
+++++++
+* Preview release
+* Tab completion for providers using index.json
+* Additional arg validation
+* az tc provider deploy: Open setup url in browser after deployment
+
 0.4.1
 ++++++
 * Preview release